+++ src/main/scala/gameover/fwk/ai/CollisionDetector.scala
@@ -14,7 +14,6 @@ trait CollisionDetector {
    * Check a position status
    */
   def checkPosition(x: Float, y: Float) : CollisionState.Value
-
   /**
    * This method check if area is intersecting something it shouldn't and return an array of
    * Rectangle objects where the collision append. The velocity of the sprite is updated accordingly.
@@ -22,25 +21,36 @@ trait CollisionDetector {
   def checkCollision(area: Rectangle, movingSpriteVelocity: Vector2, onlyBlocking: Boolean): GdxArray[Rectangle]
 
   /**
-    * Check if a point has a direct view to a target point.
+    * Check if a point has a direct view to a target point. Only blocking can block this view.
     */
-  def isDirect(x: Float, y: Float, targetX: Float, targetY: Float, onlyBlocking: Boolean): Boolean
+  def hasDirectView(x: Float, y: Float, targetX: Float, targetY: Float): Boolean
 
   /**
-   * Check if a sprite defined by its area has a direct view to a target point.
+   * Check if a sprite defined by its area has a direct view to a target point. Only blocking can block this view.
    */
-  def isDirect(visionArea: Rectangle, targetX: Float, targetY: Float, onlyBlocking: Boolean): Boolean
+  def hasDirectView(visionArea: Rectangle, targetX: Float, targetY: Float): Boolean
 
   /**
-    * Check if a sprite defined by its area has a direct view to a target area.
+    * Check if a sprite defined by its area has a direct view to a target area. Only blocking can block this view.
     */
-  def isDirect(visionArea: Rectangle, targetArea: Rectangle, onlyBlocking: Boolean): Boolean
+  def hasDirectView(visionArea: Rectangle, targetArea: Rectangle): Boolean
 
+  /**
+    * Check if a sprite can move straight to a point.
+    */
+  def canMoveStraightToPoint(area: Rectangle, targetX: Float, targetY: Float): Boolean
+
+  /**
+    * Check area vs the collision map to see if something collides.
+    */
   def checkCollision(area: Rectangle, onlyBlocking: Boolean) : Boolean = {
     val state: CollisionState.Value = checkPosition(area)
     (onlyBlocking && state == CollisionState.Blocking) || (!onlyBlocking && state != CollisionState.Empty)
   }
 
+  /**
+    * Check a point vs the collision map to see if something collides.
+    */
   def checkCollision(x: Float, y: Float, onlyBlocking: Boolean) : Boolean = {
     val state: CollisionState.Value = checkPosition(x, y)
     (onlyBlocking && state == CollisionState.Blocking) || (!onlyBlocking && state != CollisionState.Empty)